Edit movie details activity

+++ app/src/main/java/com/csovan/themoviedb/data/model/movie/Movie.java
@@ -31,17 +31,17 @@ public class Movie {
     private String backdropPath;
 
     @SerializedName("budget")
-    private Integer budget;
+    private int budget;
 
     @SerializedName("revenue")
-    private Integer revenue;
+    private long revenue;
 
     @SerializedName("genres")
     private List<MovieGenres> genres;
 
     public Movie(Integer id, String title, String releaseDate, Integer runtime, Double voteAverage,
-                 String overview, String posterPath, String backdropPath, Integer budget,
-                 Integer revenue, List<MovieGenres> genres) {
+                 String overview, String posterPath, String backdropPath,
+                 int budget, long revenue, List<MovieGenres> genres) {
         this.id = id;
         this.title = title;
         this.releaseDate = releaseDate;
@@ -119,19 +119,19 @@ public class Movie {
         this.backdropPath = backdropPath;
     }
 
-    public Integer getBudget() {
+    public int getBudget() {
         return budget;
     }
 
-    public void setBudget(Integer budget) {
+    public void setBudget(int budget) {
         this.budget = budget;
     }
 
-    public Integer getRevenue() {
+    public long getRevenue() {
         return revenue;
     }
 
-    public void setRevenue(Integer revenue) {
+    public void setRevenue(long revenue) {
         this.revenue = revenue;
     }
